APG-273 Move PNI into controller pkg
